chore: improve some types
